// client/src/components/mood/dashboard/trends-card/trends-bar-info/trends-bar-info.module.scss.d.ts
export type Styles = {
  'appear': string;
  'leftToRight': string;
  'trendsCard_Chart_Xaxis_Item_Info': string;
  'trendsCard_Chart_Xaxis_Item_Info_SectionHeading': string;
  'trendsCard_Chart_Xaxis_Item_Info_SectionParagraph': string;
  'trendsCard_Chart_Xaxis_Item_Info_SectionText': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;